Improve installation instructions in README and docs (#32)
# What does this PR do?


Improves installation instructions in readme and docs.

The quick start on the README is meant for single node with 4 GPUs. (ray
need not even be installed, because `ray.init` will start a cluster if
not running)

The installation guide is also updated. I have verified that the
quickstart works on a fresh instance.

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>