Doc fixes (#33)
Made a pass through existing docs to add and fix links.

Made a few minor docs changes along the way.

Also I added `NUM_GPUS` and `LOGGER` as envvars at the top of the
`run_gsm8k.sh` script. I imagine these two (and the data path) to be the
main trip-ups for people trying this for the first time (e.g., not
enough gpus, or no wandb api setup) so I wanted to make it easier to
edit.

---------

Co-authored-by: Charlie Ruan <charlieruan@berkeley.edu>