[Cleanup] Remove unused `normalize_reward` codepath in CriticModel (#51)
# What does this PR do?

We have this unused `normalize_reward` codepath in CriticModel. This PR
removes it and also improves some typing.

The original code here is from OpenRLHF. The code has been a no-op at
the moment - this is because the mean and std buffers are intiialized to
zero and 1, and used as-is during training.

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>